Only make one PR after release
Previously 2 commits would be made on different branches from main,
and a PR created against main on GitHub. Merging either of those would
cause a conflict in the other. Also, there's a specific order that
those commits need to merged in, so having two pull requests increases
the potential for human error.

The intention of this change is to make it so that there will only be
one PR, with the two commits ordered as they should be.