Support @BeforeEach and test cases referring to same cluster
Why:
We might want to extract some common setup to an @BeforeEach
annotated method and have it refer to the same cluster as is
used later in the test. For example to create some topics or
assert some pre-conditions like cluster size for all cases.

Note this changes an existing behaviour where it was not allowed
to have multiple KafkaClusters injected using the same name. This
makes sense because a KafkaCluster declaration also implies the
constraints required and the multiple declarations could have different
constraints. So with the new change a parameter is allowed to refer
to a cluster instantiated earlier, but not supply any constraints.