Use a new testcontainers DockerClient for every call
Why:
When used with podman we are struggling with the fact that the podman service
by default kills idle connections after 10 seconds and our HTTP requests fail
when trying to reuse pooled connections after they have idled more than 10 seconds.

This workaround creates a new client every time we do docker client work so that
a new HttpClient will be created with a new pool.