Wait for kafka cluster shutdown before tempdir clean
Why:
If kafka is running while the files are being deleted from under it
this can result in Kafka halting/exiting the process as it detects
horrifying action taking place. There is a separate PR to ensure a
halt/exit doesn't kill tests and cause baffling test failures where
the VM exits with code 1 and emits minimal logs.