Allow controller-only nodes in CONTAINER mode
Disable automatic configuration of the kafka server by the kafka-native
application, instead using the properties file exclusively for server
configuration. We also have to add log.dir to the properties file as
disabling automatic configuration prevents it from obtaining the override
from the environment.

Why:
https://github.com/ozangunalp/kafka-native/issues/88 was merged and
the docker image should now support running in controller only mode.

Note:
Had to move where we determine the roles in KafkaClusterConfig#ConfigHolder
because the process.roles are inserted in the map after ConfigHolder
construction. Needs cleaning with fire, but is too big to attack as part
of this change.