Refactor modelling of Kafka Topology
Introduces:
1. KafkaDriver: this is the connection to the InVM and Testcontainers
implementations that we use to create and destroy kafka nodes and zookeepers.
It also supplies endpoint configuration.
2. KafkaOrchestrator to create a topology and handle some higher level
operations like stopping nodes in the right order. This is where we connect
the users desired configuration with a driver.
3. KafkaTopology: this represents the (mutable) set of kafka nodes (and
optionally zookeeper) and tracks their current state. It is dynamic and
is used to stop/start/remove nodes.
4. KafkaNodeConfiguration: this represents the runtime configuration of
a node. Some of the configuration will be pulled from the user specified
KafkaClusterConfig and some will be obtained from the driver (endpoints)
5. KafkaNode: this represents the deployed node. This is an abstraction
that is implemented by each driver. It can be started/stopped/removed.
6. ZookeeperConfiguration: as above this is the runtime configuration of
zookeeper. Currently just it's endpoint details obtained from the driver.
7. Zookeeper: as above, this is an abstraction of the zookeeper process
implemented by each each driver.

KafkaNodeConfiguration is responsible for generating the properties for
the server using it's knowledge of what roles the node has. The responsibilities
of ConfigHolder are now owned by KafkaNodeConfiguration (making certain
ports and endpoint addresses accessible). It can consult it's parent
topology to discover information about the other nodes (for example to
build up the quorum address list).

Why:
We've noticed in several PRs that it feels bad usings the nodeId
combined with the user's desired quantity of controllers and brokers
to try and determine the roles of the node. This along with the tangle
of maps made it tricky to reason about. For instance, recently when
enabling controller only mode we had to be careful where we inferred
from the generated propertieds that the node was a controller. This
change makes the topology of the cluster a part of the model so that
we can designate it's roles and other characteristics up front. We
won't need to look at it's properties to determine if it's a broker
or controller because it will be an immutable property of that node.