PR Feedback. Custom ImagePullPolicy
Enforce we always pull Release and Snapshot tags. It then works of the assumption that older versioned tags are stable so the cached copy is likely to be good.

It will fall back to the default policy if the versioned image is less than 7 days old.

Signed-off-by: Sam Barker <sbarker@redhat.com>

rh-pre-commit.version: 2.0.1
rh-pre-commit.check-secrets: ENABLED