Adopt apache kafka container images for 4.1+
If the target version is greater than or equal to 4.1, or the version is
'latest', then we use the apache/kafka docker images.

Unlike the images we used prior, the apache images do not have a convenience
for setting up SCRAM users at storage formatting time via env var, instead
we use an admin client to create the users after the server starts via the
anonymous listener. Using the public APIs is a better approach anyway.

We currently do not use the apache kafka-native images because there are
some problems with SASL authentication, see:
- https://issues.apache.org/jira/browse/KAFKA-19583
- https://issues.apache.org/jira/browse/KAFKA-19584

Signed-off-by: Robert Young <robertyoungnz@gmail.com>