add TopicNamingStrategy enum to control auto topic naming
The extension can create and automatically name topics for users, with
the theory that users often just want a topic to operate on. If the user
has to be responsible for the naming it's often just irrelevant noise.
So the framework handles it.

However, sometimes the topic naming may have other consequences. For
example if some technology wants to use the topic names for some other
purpose, it can be useful to constrain the value. For instance so that
the topic name also happens to be a valid identifier in a third party
system (the example I'm thinking of is that it is a valid Key name in a
cloud provided KMS).

The two strategies are:
* RANDOM_ADJECTIVE_UNDERSCORE_NOUN - eg unruffled_hawking
* RANDOM_ADJECTIVE_HYPHEN_NOUN - eg suspicious-newton

For example to control the naming of a parameter injected Topic:
```java
@TopicNamingStrategy(NamingStrategy.RANDOM_ADJECTIVE_HYPHEN_NOUN) Topic topic
```

Signed-off-by: Robert Young <robertyoungnz@gmail.com>